% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportRPATMetrics.R
\name{adjustHvyVehAgeDistribution}
\alias{adjustHvyVehAgeDistribution}
\title{Adjust cumulative age distribution to match target ratio for heavy vehicles}
\usage{
adjustHvyVehAgeDistribution(CumDist_, AdjRatio)
}
\arguments{
\item{CumDist_}{A named numeric vector where the names are vehicle ages and
the values are the proportion of vehicles that age or younger. The names must
be an ordered sequence from 0 to 32.}

\item{AdjRatio}{A number that is the target ratio value.}
}
\value{
A numeric vector of adjusted distribution.
}
\description{
\code{adjustHvyVehAgeDistribution} Adjusts a cumulative age distribution to match a
target ratio for heavy vehicles.
}
\details{
This function adjusts a cumulative age distribution to match a target ratio.
The function returns the adjusted cumulative age distribution and the
corresponding age distribution.
}
