% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportRPATMetrics.R
\name{assignHvyVehFuelEconomy}
\alias{assignHvyVehFuelEconomy}
\title{Assignes mileage to heavy vehicles}
\usage{
assignHvyVehFuelEconomy(AgeDist_Ag, Mpg__Yr = TrkBusMpg__Yr, Type,
  CurrYear)
}
\arguments{
\item{AgeDist_Ag}{Age distribution of vehicle type.}

\item{Mpg__Yr}{A data frame of mileage of vehicles by type and year.}

\item{Type}{A string identifying the type of vehicle ("Truck" or "Bus").}

\item{CurrentYear}{A integer indicating the current year.}
}
\value{
A numeric vector that indicates the mileage of vehicles.
}
\description{
\code{assignHvyVehFuelEconomy} Assignes mileage to heavy vehicles.
}
\details{
This function assigns mileage to heavy vehicles like truck or bus.
}
