% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDatastore.R
\name{documentDatastoreTables}
\alias{documentDatastoreTables}
\title{Save an zip archive which documents all tables/datasets in a datastore.}
\usage{
documentDatastoreTables(SaveArchiveName, QueryPrep_ls)
}
\arguments{
\item{SaveArchiveName}{a string identifying the name of the zip archive file
that will be saved. The name should not include any suffix (e.g. '.zip') as
that will be automatically added. The file will be saved in the current
working directory.}

\item{QueryPrep_ls}{a list created by calling the prepareForDatastoreQuery
function which identifies the datastore location(s), listing(s), and
functions for listing and read the datastore(s).}
}
\value{
a logical identifying whether the archive file has been saved.
}
\description{
\code{documentDatastoreTables} saves a zip archive of a set of csv-formatted
text files which document tables in a datastore.
}
\details{
This function inventories all datsets in the tables in a datastore and
creates a set of csv-formatted text files where the dataset inventory for
each table is in a csv-formatted text file having the name of the table.
Each csv file lists the names of the datasets, their types, their units, and
descriptions. The files are saved in a zip archive organized by datastore
group.
}
