% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDatastore.R
\name{documentGroupDatasets}
\alias{documentGroupDatasets}
\title{Save an Excel workbook to document all tables/datasets in a datastore group.}
\usage{
documentGroupDatasets(Group, Filename, QueryPrep_ls)
}
\arguments{
\item{QueryPrep_ls}{a list created by calling the prepareForDatastoreQuery
function which identifies the datastore location(s), listing(s), and
functions for listing and read the datastore(s).}

\item{a}{string identifying the name of the Group in the datastore.}

\item{a}{string identifying the name of the Excel workbook file to save. The
name must include the relative path information if it is to be saved in other
than the current working directory.}
}
\value{
a logical identifying whether the Excel file has been saved.
}
\description{
\code{documentGroupDatasets} save an Excel workbook which documents
datasets in all tables in a datastore group.
}
\details{
This function inventories all datsets in the tables in a datastore group and
creates an Excel workbook where the dataset inventory for each table is in
a separate worksheet. Each worksheet lists the names of the datasets, their
types, their units, and descriptions.
}
