% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDatastore.R
\name{listDatasets}
\alias{listDatasets}
\title{List names and descriptions datasets in a table in a datastore.}
\usage{
listDatasets(Table, Group, QueryPrep_ls)
}
\arguments{
\item{Table}{a string that is the name of the table to retrieve the dataset
names and descriptions from.}

\item{Group}{a string that is the name of the group to retrieve the table
datasets from.}

\item{QueryPrep_ls}{a list created by calling the prepareForDatastoreQuery
function which identifies the datastore location(s), listing(s), and
functions for listing and read the datastore(s).}
}
\value{
A data frame which lists the dataset names and descriptions for
datasets in the identified table.
}
\description{
\code{listTables} a function which lists the names and descriptions datasets
in a table in a datastore.
}
\details{
This functions lists the names and descriptions datasets in a table in a
datastore.
}
