% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDatastore.R
\name{listGroups}
\alias{listGroups}
\title{Lists the names of groups in model datastores.}
\usage{
listGroups(QueryPrep_ls)
}
\arguments{
\item{QueryPrep_ls}{a list created by calling the prepareForDatastoreQuery
function which identifies the datastore location(s), listing(s), and
functions for listing and read the datastore(s).}
}
\value{
A named list where each component is a vector of group names for a
datastore.
}
\description{
\code{listGroups} a function which lists the groups in a datastore
or datastores that contain data for a scenario.
}
\details{
This function lists the names of groups in a model datastore and any other
datastores that are referenced by a model run.
}
