% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDatastore.R
\name{listTables}
\alias{listTables}
\title{List names of tables in a group in a datastore.}
\usage{
listTables(Group, QueryPrep_ls)
}
\arguments{
\item{Group}{a string that is the name of the group to retrieve the table
names from.}

\item{QueryPrep_ls}{a list created by calling the prepareForDatastoreQuery
function which identifies the datastore location(s), listing(s), and
functions for listing and read the datastore(s).}
}
\value{
A named list where each component is a vector of table names in the
group in the datastore(s).
}
\description{
\code{listTables} a function which lists the tables in a group in a
datastore.
}
\details{
This functions lists the tables in a group in a datastore.
}
