% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryDatastore.R
\name{prepareForDatastoreQuery}
\alias{prepareForDatastoreQuery}
\title{Retrieve information needed to prepare a datastore query.}
\usage{
prepareForDatastoreQuery(DstoreLocs_, DstoreType)
}
\arguments{
\item{DstoreLocs_}{a string vector identifying the paths to all of the
datastores to extract the datasets from. Each entry must be the full relative
path to a datastore (e.g. 'tests/Datastore').}

\item{DstoreType}{a string identifying the type of datastore
(e.g. 'RD', 'H5').}
}
\value{
A named list having three components. The 'Dir' component is a
string vector identifying the relative path(s) to the datastore(s). The
'Listing' component is a list where each component is a datastore listing.
The 'Functions' component contains the appropriate functions for the
datastore type for listing the datastore contents and for reading datasets.
}
\description{
\code{prepareForDatastoreQuery} retrieves datastore listings and functions
required to make a datastore query.
}
\details{
This function prepares a list of information that is used in making a query
of a VisionEval model datastore. The list includes the location(s) of the
datastore(s) to be queried, the listing(s) for those datastores, and the
functions to be used for reading the datastore(s). More than one datastore
may be specified so that if datastore references are used in a model run,
datasets from the referenced datastores may be queried as well. Note that the
capability for querying multiple datastores is only for the purpose of
querying datastores for a single model scenario. This capability should not
be used to compare multiple scenarios. The function does not segregate
datasets by datastore. Attempting to use this function to compare multiple
scenarios could produce unpredictable results.
}
