% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportRPATMetrics.R
\name{ReportRPATMetrics}
\alias{ReportRPATMetrics}
\title{Function to calculate performance metrics.}
\usage{
ReportRPATMetrics(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{ReportRPATMetrics} calculates performance metrics.
}
\details{
This function calculates and tabulates various performance metrics
that ranges from environmental and energy impacts, financial and
economic impacts, and community impacts.
}
