% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportRPATMetrics.R
\docType{data}
\name{TruckBusAgeDist_mx}
\alias{TruckBusAgeDist_mx}
\title{Truck and Bus age distribution data}
\format{A matrix with two columns containing:
\describe{
  \item{Truck}{age distribution for trucks}
  \item{Bus}{age distribution for buses}
}}
\source{
CalculateTravelDemand.R script.
}
\usage{
TruckBusAgeDist_mx
}
\description{
A matrix of age distribution for trucks and buses
}
\keyword{datasets}
